+++ index.html
@@ -230,6 +230,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
+  <script type="module" src="./js/data.js"></script>
 </body>
 </html>
